+++ src/types/comment.ts
@@ -1,0 +1,10 @@
+import { Timestamp } from 'firebase/firestore';
+
+export interface CommentModel {
+  id: string;
+  userId: string;
+  content: string;
+  createdAt: Timestamp;
+  updatedAt?: Timestamp;
+  parentCommentId?: string;
+}
